+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Chore
 
+- [#7781](https://github.com/blockscout/blockscout/pull/7781) - Add `/api/v1/health/liveness` and `/api/v1/health/readiness`
+
 <details>
   <summary>Dependencies version bumps</summary>